LayoutTests/fast/css-grid-layout/flex-and-intrinsic-sizes.html

<!DOCTYPE html>

<link href="../css-intrinsic-dimensions/resources/width-keyword-classes.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/height-keyword-classes.css" rel="stylesheet">
<link href="resources/grid.css" rel="stylesheet">

<style>
.container {
    width: 100px;
    height: 100px;
}

.grid {
  -webkit-grid-template-columns: 1fr;
  -webkit-grid-template-rows: 1fr;
}

div { font: 10px/1 Ahem; }

</style>

<script src="../../resources/check-layout.js"></script>

<body onload="checkLayout('.grid')">

<p>This test checks that fr tracks are properly sized whenever grid have intrinsic sizes.</p>

<div class="container">
  <div class="grid" data-expected-width="100" data-expected-height="10">
    <div>XXX XXX</div>
  </div>
</div>

<!-- This fails due to https://bugs.webkit.org/show_bug.cgi?id=150674 -->
<div class="container">
  <div class="grid min-content" data-expected-width="30" data-expected-height="10">
    <div>XXX XXX</div>
  </div>
</div>

<div class="container">
  <div class="grid max-content" data-expected-width="70" data-expected-height="10">
    <div>XXX XXX</div>
  </div>
</div>

<div class="container">
  <div class="grid fit-content" data-expected-width="70" data-expected-height="10">
    <div>XXX XXX</div>
  </div>
</div>

<div class="container">
  <div class="grid fill-available" data-expected-width="100" data-expected-height="100">
    <div>XXX XXX</div>
  </div>
</div>

</body>
